Refactor: 이메일과 소셜 정보 컴포넌트 분리
